requirements.txt
Flask==3.1.1
python-dotenv==1.1.0
psutil==5.9.5
pinecone-client==4.1.0
langchain==1.0.5
langchain-core==1.0.4
langchain-groq==1.0.0
groq==0.33.0
langchain-huggingface==1.0.1
huggingface-hub==0.33.4
sentence-transformers==2.2.2
langchain-community==0.4.1
pypdf==5.6.1
tqdm==4.67.1
